+++ examples/data-sources/anytype_property/data-source.tf
@@ -1,0 +1,8 @@
+data "anytype_property" "status" {
+  space_id = "bafyreigyfkt6rbv24sbv5aq2hko3bhmv5xxlf22b4bypdu6j7hnphm3psq.23me69r569oi1"
+  id       = "bafyreids36kpw5ppuwm3ce2p4ezb3ab7cihhkq6yfbwzwpp4mln7rcgw7a"
+}
+
+output "status_format" {
+  value = data.anytype_property.status.format
+}