+++ internal/provider/provider.go
@@ -101,6 +101,8 @@ func (p *AnytypeProvider) Configure(ctx context.Context, req provider.ConfigureR
 func (p *AnytypeProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewSpaceResource,
+		NewTypeResource,
+		NewPropertyResource,
 	}
 }
 
@@ -108,6 +110,10 @@ func (p *AnytypeProvider) DataSources(_ context.Context) []func() datasource.Dat
 	return []func() datasource.DataSource{
 		NewSpaceDataSource,
 		NewSpacesDataSource,
+		NewTypeDataSource,
+		NewTypesDataSource,
+		NewPropertyDataSource,
+		NewPropertiesDataSource,
 	}
 }